requirements.txt
streamlit==1.32.0
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.8.1.78
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
pydicom==2.4.3
seaborn==0.13.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
python-dotenv==1.0.0
